[writeboost] Add rambuf_read_fullsize test
RAM buffer is really short-lived - it is soon flushed to cache device.
Thus, read-hit on RAM buffer is really unlikely in real-world system
which enables page cache to read in upper layer than block.

This test exploits PatternStomper to cover the RAM-buffer-read-hit case.
Also, WriteboostStatus is used to see if it really experienced the path.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>