[writeboost] Sync daemon works while drop_caches
drop_caches waits for all dirty caches written back to the backing
device. If some dirties stores after this message is sent, it never
returns (it is interruptible, the kernel code is OK).
To avoid this failure, we need to work the sync daemon while
drop_caches. This periodically flushes the dirties on the RAM buffer so
this failure never happen.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>