[writeboost] Remove barrier_deadline_ms
In writeboost, Processing of write barriers is deferred.

Now there are two stages, 1) tunable deferment by timer which
queues a work and then 2) the work is dispatched.

This patch removes the first stage.

dm-thin and dm-cache do almost the same thing but they don't have such
complicated mechanism but only the second stage.
As a result of having the first stage, the code is complicated and hard
to understand, and also complicates the user interface.
I think of removing of this tunable for this reason.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>