[lexical_operators] Rename create method and make it public
This allows reusing the common device creation code in places where
using the with_dev* context managers is not suitable.

This is needed by the dm-snapshot test suite in places where we want to
create and activate snapshots which persist beyond the scope of a
context manager, e.g., across loop iterations.

Signed-off-by: Nikos Tsironis <ntsironis@arrikto.com>