[blktrace] Run blkparse using IO.popen
Until now we ran blkparse using backticks. This method stores the whole
output of blkparse in memory before we get to process it. This is
inefficient both memory and performance-wise.

Run blkparse with IO.popen so we can process its output incrementally
through a pipe.

In a test, where we processed 4.5G of blktrace traces, this change
reduced Ruby's memory usage from 12G to 3.7G and the test's run time
from 10m19.409s to 5m31.387s.

Signed-off-by: Nikos Tsironis <ntsironis@arrikto.com>