[blktrace] Fix Event class
Commit 68e0383ae3b4 ("[blktrace] Introduce CPUIODistribution class")
extended the Event class to include the CPU id of the CPU that performed
an I/O operation.

This change breaks tests, e.g., the thin-provisioning
FlushTriggersCommitTests, that create an Event() object describing a
particular I/O operation and check if it's a member of the Event list
returned by blktrace().

Customize the Event class by overloading the == operator to ignore the
cpu field, if for either of the Event objects the field is nil.

This allows us to search for particular events, in the Event list
returned by blktrace(), without specifying the CPU of the event.

Fixes 68e0383ae3b4 ("[blktrace] Introduce CPUIODistribution class")
Signed-off-by: Nikos Tsironis <ntsironis@arrikto.com>