[cache] Disable discard passdown in the discard_with_concurrent_io test
The test was written prior to the discard passdown feature was
introduced in kernel v5.1 and is not compatible to the passdown
feature. Passing down discards to the origin in this test leads
to data errors because written but uncached blocks get trimmed
unexpectedly before read verification. Set the cache target
no_discard_passdown to avoid test failures.